Add RepositoryPostgresTestBase
